backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
pandas
numpy
nltk
python-dotenv
